refactor(api): split validateEmail from user service
