+++ packages/api/src/modules/user/services/check/user.check.service.test.ts
@@ -1,4 +1,9 @@
 import userCheckService from "./user.check.service";
+import { UserServiceErrors } from "../../user.service";
+import configurationsService from "../../../configurations/configurations.service";
+import { BadRequestError } from "../../../../shared/errors/httpErrors";
+jest.mock("../../../configurations/configurations.service");
+const mockedConfigurationsService = jest.mocked(configurationsService);
 
 describe("user check service", () => {
     describe("passwordValidator", () => {
@@ -37,6 +42,35 @@ describe("user check service", () => {
 
         it("should accept", () => {
             expect(userCheckService.passwordValidator("SUPER;test::12345678")).toBe(true);
+        });
+    });
+
+    describe("validateEmail()", () => {
+        mockedConfigurationsService.isDomainAccepted.mockImplementation(async () => true);
+        const EMAIL = "daemon.targaryen@ac-pentos.ws";
+
+        it("should verify domain", async () => {
+            await userCheckService.validateEmail(EMAIL);
+            expect(mockedConfigurationsService.isDomainAccepted).toHaveBeenCalledWith(EMAIL);
+        });
+
+        it("should return if email is correct", async () => {
+            await userCheckService.validateEmail(EMAIL);
+        });
+
+        it("should throw error if not well formatted", async () => {
+            const expected = new BadRequestError("Email is not valid");
+            expect(() => userCheckService.validateEmail("ab1")).rejects.toThrowError(expected);
+        });
+
+        it("should throw error if domain not accepted", async () => {
+            mockedConfigurationsService.isDomainAccepted.mockImplementationOnce(async () => false);
+            const expected = {
+                message: "Email domain is not accepted",
+                code: UserServiceErrors.CREATE_EMAIL_GOUV,
+            };
+            const test = () => userCheckService.validateEmail(EMAIL);
+            await expect(test).rejects.toMatchObject(expected);
         });
     });
 });
